Update arrayifs.bas
Moved from Boolean argument "asstring" that sets the choice between string result or array,  to string argument "delimiter". Its default value is "" and in that case the result of the function is an array. If it is set to any character, like "," or " ", then the result would be string delimited using that same argument.